Fix module loading and release v0.1.2
One can specify an npm module name to load in the configuration and/or
specification file. Node then tries to find it in the usual way - by
searching the path up to root for that module and in the global modules'
directory. However, node starts the search from hyperswitch' path,
not the application's. Unfortunately, this doesn't work in cases where
node modules are symlinked into the relevant node_modules directory. So,
when a relative path is encountered and the module fails to load, try to
load it both from the application's base path as well as its
node_modules sub-directory, if one exists.