Make sure that default route filters are processed first (#46)
We generally want to record metrics for all requests, even those that are
rejected by other filters. For this reason, it is important that the metrics
filter is installed highest in the stack.

This patch generally changes the initialization so that default filters are
placed highest in the stack, followed by custom filters from the spec.