Fix specRoot / basePath handling
Don't rely on specRoot.basePath to match the current domain, and instead just
check for the presence of specRoot. As we now configure specRoot at the root
of the hierarchy as well, we do not need the exception for the domain listing
at `/` any more.

Task: https://phabricator.wikimedia.org/T142366