Fix swagger option mangling
The Sorter option we previously relied on for a part of our swagger UI
option mangling pass was removed. This meant that some of the options we
used to set were no longer making it into the HTML.

This patch moves those options to another replace, which does exist.
Among other things, this fixes
https://phabricator.wikimedia.org/T170729.

Bug: https://phabricator.wikimedia.org/T170729