Deprecate the use of the options.log or hyper.log
By exposing the bound log functions like options.log or hyper.log
instead of forcing clients to use the hyper.logger object directly,
we mask a powerful ability to add childer loggers to the exposed
logger from clients. Also, hyper itself can not augument the logger
on the per-request basis, for example to add the URI spec to the request
after the match succeeded. All clients should only use hyper.logger
to log.

Current change is backwards compatible, but the fallbacks will be removed
as soon as restbase and change-prop adapts to the new logging method.