Have proper module caching and sharing
Module caching and sharing has always been completely broken. That
stemmed  from the fact that using the `===` operator on Objects results
in a truth-y evaluation iff the objects being tested have the same
reference, i.e. they are the same object. Objects were used both as keys
and as values during the caching process, which meant that no module was
ever shared effectively. Furthermore, it is important that the context
in which the module is being loaded matches the one used by the
previous-cached module. Here, again, the process suffered from equality
comparison problems.

In order to make module caching and sharing work, this commit switches
to using the SHA1 digest of the module definition as its caching key,
allowing us to retrieve the cached module based on the value of the
definition instead of its reference. To counter the problem of context
checking, we now cache an 'exportable' version of the context as well
in addition to the module definition.