Allow the service to skip issuing resource requests
A module may specify PUT requests to be made once the loading of the
spec is complete in order to fully set the system up. However, in the
case of RESTRouter that is not needed nor desirable, since (i) RESTBase
back-end already performs these; and (ii) issuing these requests over
network significantly prolongs the start-up sequence.